==> ATraiU 2.0/code/1_Organizing_physio_data.Rmd <==
---
title: "Organizing Results of Physiological Literature Search"
author: "TP DuBose"
date: "07/01/2022"
output: pdf_document
---

```{r setup, include=F}
library(tidyverse);library(googlesheets4); library(googledrive);library(knitr)
```

# Bring in the results of our concentrated physiological search

First, I use the *googlesheets4* and *googledrive* packages to query our shared google drive for the results of our concentrated literature search (completed mostly by Vj in 2020-2021). You might need to go to the console to complete the drive_find and read_sheet function.

```{r}
# identify sheets from concentraited literature search
sp_files<-drive_find(shared_drive = "Anuran Physiological Trait Search", 
           pattern="MNM traits")
# pull those sheets in to R
all_trait_shts<-NULL
for(i in 1:nrow(sp_files)){
  file.id<-sp_files[[2]][i]
  file.taxa<-sub(" MNM traits", "", sp_files[[1]][i])
  sht<-read_sheet(file.id, col_types = 'c') %>% 
    mutate(taxa=file.taxa)
  all_trait_shts<-bind_rows(all_trait_shts,sht)
}
# write out the results
write_csv(all_trait_shts, paste0('../intermediate_results/RAW_concentrated_trait_values',
                                 format(Sys.Date(), "%Y%m%d"),'.csv'))
```

We increased the focal species in 2022. The below code adds a column for who double checked which trait values.

```{r}
all_trait_shts<-all_trait_shts %>%
  mutate(checked=ifelse(taxa %in% c('Pseudacris kalmi','Pseudacris crucifer',
                                    'Lithobates virgatipes','Lithobates sphenocephalus',
                                    'Lithobates sevosus','Lithobates okaloosae',
                                    'Lithobates kauffeldi','Lithobates heckscheri',
                                    'Incilius nebulifer','Dryophytes femoralis',
                                    'Dryophytes avivoca','Dryophytes andersonii',
                                    'Acris blanchardi'),
                        `Checked by`,'tpd/vc')) %>%
  select(-`Checked by`)
```

I then fix some data entry issues, remove empty rows or rows with sources we aren't including, and remove traits we are not interested in. Traits removed were often collected pre-2020 and were not searched for otherwise, thus we don't include them in our database. Traits that were collected but not kept include: `r unique(all_trait_shts$Trait)[c(8,11:16)]` .

```{r}
#replace N/A and n/a with NA
all_trait_shts <- all_trait_shts %>%
  mutate(across(everything()), na_if(.,"N/A")) %>%
  mutate(across(everything()), na_if(.,"n/a")) %>%
          # remove other traits
  filter(!is.na(Value), # remove empty rows
         # remove non-published sources
         `Source file name` != 'Animal Diversity Web') 
```

Then we load our tangentially collected traits. These include traits from chapter 8 and chapter 9 of Felden & Bru 199, traits listed in references found in the above taxa-focused search, and those identified within Herpetological Review 196 - 2012 issues. 

```{r}
# chapters 8 and 9 of the amphibian ecophysiology textbook
FB_sheet<-drive_find(shared_drive = "Anuran Physiological Trait Search", 
                     pattern="Parameters")
FB_trait_data<-read_sheet(FB_sheet$id, sheet="Table 9.2", col_types='c') %>%
   mutate(Trait='Tpref', Units='C') %>%
  bind_rows(read_sheet(FB_sheet$id, sheet="Ch8Text", col_types='c')) %>%
  rename(Species_orig=Taxa)
# traits found in references identified in taxa-focused search
misc_sheet<-drive_find(shared_drive = "Anuran Physiological Trait Search", 
           pattern="Miscel")
misc_trait_data<-read_sheet(misc_sheet$id, sheet='MiscFromLit', col_types='c')
# traits found within Herpetological Review
herp_rev_data<-read_sheet(misc_sheet$id, sheet='HerpReviewData', col_types='c')

# write out all of these traits
bind_rows(FB_trait_data, misc_trait_data, herp_rev_data) %>% 
  write_csv(paste0('../intermediate_results/RAW_tangental_trait_values',
                   format(Sys.Date(), "%Y%m%d"),'.csv'))
```

To consolidate these data sources and make sure we have correct taxonomic information, below I use the name_backbone function to identify valid taxon for each recorded species names. 

```{r taxonomy check}
library(rgbif)
anura_tax<-unique(c(herp_rev_data$Species_orig, misc_trait_data$Species_orig, FB_trait_data$Species_orig, all_trait_shts$taxa))
anuran.gbif.taxa<-NULL
for(u in anura_tax){
  anuran.gbif.taxa1<-name_backbone(name=u, 
                                   rank='species', kingdom='animals')
  anuran.gbif.taxa<-bind_rows(anuran.gbif.taxa1 %>% mutate(dataScientificName=u), 
                              anuran.gbif.taxa)
}
taxa_key_start<- anuran.gbif.taxa %>%
  select(dataScientificName, canonicalName, scientificName, family, genus, species, usageKey)
write.csv(taxa_key_start, '../intermediate_results/taxonomic_key_start.csv', row.names = F)
```

I then checked taxa for spelling mistakes on 01/20/2022 and 05/12/2022. I also edited the table to to link stray taxa to new names using ITIS, i.e. Hyla raddiana. Below is the function I used to query stray taxa and bring in the updated taxonomic list.

```{r}
name_backbone(name='Boana riojana', rank='species', kingdom = 'animals') %>%
  select(canonicalName, scientificName, family, genus, species, usageKey)
taxa_key <- read.csv('../intermediate_results/ATraiU2_taxonomic_key_2022JULY.csv')
taxa_key_start[!(taxa_key_start$dataScientificName %in% taxa_key$dataScientificName),]
#taxa_key[taxa_key$dataScientificNam!=taxa_key$canonicalName,]
```

# Build Reference List ----

I used the below code to identify references with doi and cross reference to populate author and journal information. I haven't used this code chunk since July 7, 2022 though. I looked up missing references by hand. I also collected all the reference information from the Herpetelogical Review.

```{r, eval=F}
#remotes::install_github("ropensci/rcrossref")
library(rcrossref)
test<-all_trait_shts$DOI[1]
cr_cn(test,format='text')
rm(test)
doi.refs<-all_trait_shts %>% 
  rename(`Source file name`=Source.file.name) %>%
  filter(!is.na(DOI),
         !(DOI %in% c("NULL", "-", 28)),
         !grepl('Rausch', `Source file name`)) %>%
  distinct(`Source file name`, DOI) %>%
  rowwise() %>% 
  mutate(citation=list(cr_cn(dois=sub("DOI ", "", DOI), format='text')))
#will get lots of blue text, just need to hand fix Rausch
#49 is bad
doi.refs<-doi.refs %>%
  mutate(citation=ifelse(is.null(citation),NA,unlist(citation)))

write_csv(doi.refs, paste0('../intermediate_results/RAW_references_', format(Sys.Date(), "%Y%m%d"), '.csv'), )
```

After entering all the references, I add a unique ID for each reference in the code chunk below. 

```{r, eval=F}
#bring the reference data in
refs<-read.csv('../ATraiU2_reference_list_2022SEPT.csv') %>% 
  group_by(citation) %>% 
  mutate(group_id=cur_group_id()) %>%
  rename(refID=group_id) %>%
  ungroup()
write.csv(refs, '../ATraiU2_reference_list_2022SEPT.csv', row.names = F)
```
```{r print refs, echo=F}
refs<-read.csv('../ATraiU2_reference_list_2022SEPT.csv')
kable(refs %>% slice_sample(n=4))
```

# Collate Data

This code combines all the trait information into one datasheet. I changed the collection method for species identified as *Pseudacris maculata*, which is not found within the southeastern United States. It was originally identified as *Pseudacris triseriata* but, after using Lemon et al. 2007 to check the taxonomic identities associated with *P. triseriata* trait values, none remained.

```{r}
trait_db_raw<-all_trait_shts %>%
  rename(dataScientificName=taxa) %>%
  # Pseudacris maculata doesn't occur within the southeast 
  mutate(dataCollationType=ifelse(dataScientificName == 'Pseudacris maculata',
                                  'haphazard', 'concentrated')) %>%
  # bind with tangental traits
  bind_rows(bind_rows(FB_trait_data, misc_trait_data, herp_rev_data) %>%
              rename(dataScientificName=Species_orig) %>% 
              mutate(dataCollationType='haphazard')) %>%
  left_join(taxa_key, by='dataScientificName') %>%
  # clean up the text
  mutate(Trait=recode(Trait, "Standard Metabolic Rate"="Metabolic Rate",
                      "Cutaneous Water Loss Rate"="Water Loss Rate"),
         # fix misspellings
         `Life Stage`=recode(`Life Stage`, 'juveniles' = 'juvenile', 
                             'unkown'='unknown', 'tapdole'='tadpole')) %>%
  select(dataScientificName, species, Trait, everything()) # reorders columns 
write_csv(trait_db_raw, '../intermediate_results/combined_trait_data.csv')
#trait_db_raw<-read.csv('../intermediate_results/combined_trait_data.csv')
```

This also gives me a chance to geolocate species whose identity might have changed since publication (i.e., Acris and Psuedacris).

```{r, echo=F}
kable(trait_db_raw %>%
  filter(genus == 'Pseudacris',
         !grepl('Hyla ', dataScientificName),
         !grepl('crucifer', dataScientificName)) %>%
    distinct(dataScientificName, Location,`Source file name`, Notes)) 
kable(trait_db_raw %>%
  filter(genus == 'Acris', `Source file name` !="AmphibiaWeb") %>%
  distinct(dataScientificName, Location,`Source file name`))
```
# Denormalized Trait Database

Then I created a denormalized trait database of all the data. This contains all the information, where each row is a trait value and columns correspond to different attributes associated with that trait value. Some of those columns are superfluous or repetitive.

```{r}
trait_db_raw %>%
  arrange(dataCollationType) %>%
  #remove traits that have been duplicated on accident
  distinct(species, Trait, `Life Stage`, 
           Sex, Value, Units, Location, Acclimation, .keep_all=T) %>%
  left_join(refs, by=c(`Source file name`="traitSource")) %>% 
  rename(traitName = Trait, traitValue = Value, traitUnit = Units,
          lifeStage=`Life Stage`, sex=Sex, verbatimLocality=Location) %>%
   #filter(is.na(first_author)) %>% pull(`Source file name`)
  write_csv('../intermediate_results/denormalized_trait_db.csv')
```

# Final Collation - Database Files Produced Here

In the following tables, I often change column names to match ecological trait standards when possible. These standards are outlined at: [Ecological Trait-data Standard](https://ecologicaltraitdata.github.io/ETS/). First, we create a long dataframe, were each row is a unique trait value. This does not contain the superfluous columns that the denormalized trait database might have. We also reduce the species pool to only those that occur within the southeastern United States (our 40 focal species). 

```{r}
trait_long<-trait_db_raw %>%
  rename(traitSource=Source.file.name) %>%
  # fixing spelling errors in references to consolidate to one traitSource
  mutate(traitSource=case_when(traitSource=="Dupre and Petranka 1985" ~ "DuprePetranka1985_Tpref",
                               traitSource=="Fitch1956" ~ "Fitch1956_temperature",
                               T ~ traitSource)) %>%
  left_join(refs, by='traitSource') %>% 
  # remove columns that are superfluous for larger tables
  select(-dataScientificName, 
         -`Search.Terms.Used`, -DOI, -page, -Season,
         -doi, -citation, -journal, -issue, -year,
         -ends_with('author'), -`Original.Source`) %>%
  # rename columns to match ecological trait standards when possible
  rename(traitName = Trait, traitValue = Value, traitUnit = Units,
         lifeStage=`Life.Stage`, sex=Sex, verbatimLocality=Location,
         # these just clarify & match style of above standards
         traitVariation=`Variation.of.Value`,
         traitVariationType = `Measure.of.Variation`,
         traitEvidenceCategory=`Strength.of.Evidence`,
         traitAcclimation=Acclimation) %>%
  ungroup() %>% 
  filter(!is.na(refID), # removing traits from websites
         # removing traits we did not do specific trait searches for
         !(traitName %in% c('per_surface_area', 'Thermal Perfomance Curve',
                        'Standard Metabolic Rate', 'Metabolic Rate', 'Metabolism Rate',
                        'Minimum Egg Development Temperature',
                        'Maximum Egg Development Temperature', 'Water Loss Rate'))) %>% 
  # clean up sex column
  mutate(sex=ifelse(is.na(sex), 'unknown',sex)) %>%
  # remove transcription duplicates from the database
  distinct(traitName, species, refID, traitValue, .keep_all=T) %>%
  # removing taxa not within the southeastern US
  filter(species %in% gsub(' MNM traits', '', sp_files$name),
         species != 'Pseudacris maculata', species != 'Anaxyrus punctatus')
```

Interestingly, the following ways are used to describe trait variation within studies: `r trait_long$traitVariationType %>% unique()`.

## Species x trait summary matrix

This might be used at the species level, where you just want a summary value for the entire species. They only incorporate information from adults. I separated the character traits (Activity) and convert it to a matrix of n records, where n might relay the strength of that trait (so always recorded as diurnal). These character traits are then bound the to numeric traits summarized below (where min values are summerized by the minmum, max values are summarized by the maximum, and most others are summarized by a mean). 

```{r}
#character traits
tchar<-trait_long %>%
  group_by(species, lifeStage, traitName) %>%
  filter(traitName == 'Activity') %>%
  mutate(record=1,
         activity=tolower(traitValue)) %>%
  ungroup() %>%
  dplyr::select(family, genus, species, refID, traitSource, activity, record) %>%
  pivot_wider(names_from=activity, values_from=record) %>% 
  group_by(species) %>%
  summarize(refIDs_a=paste(unique(refID), collapse=', '),
            across(c('nocturnal', 'diurnal','crepuscular','arrhythmic'), sum, na.rm=T)) %>%
  mutate(lifeStage='adult')
```
```{r, echo=F}
kable(tchar %>% slice_sample(n=3))
```
```{r}
# numeric traits
tbif<-trait_long %>%
  mutate(lifeStage = case_when(grepl('stage', lifeStage) ~ 'tadpole',
                                  lifeStage=='unknown' ~ 'adult',
                                  T ~ lifeStage)) %>%
  group_by(species, lifeStage, traitName) %>%
  # remove extra traits
  filter(traitName != 'Activity',
         lifeStage=='adult',
         traitValue != 'overwinters') %>% # to avoid warnings from as.numeric
  mutate(v_n=as.numeric(traitValue)) %>%
  select(species, lifeStage, refID, traitName, traitValue, v_n) %>%
  pivot_wider(names_from=traitName, values_from = v_n, values_fn=mean) %>%
  group_by(species, lifeStage) %>%
  summarize(refIDs_n=paste(unique(refID), collapse=', '),
            Mass=ifelse(!all(is.na(Mass)), round(mean(Mass,na.rm=T),2), NA),
            CTmax=ifelse(!all(is.na(CTmax)), max(CTmax, na.rm=T), NA),
            CTmin=ifelse(!all(is.na(CTmin)), min(CTmin, na.rm=T), NA), 
            Tpref=ifelse(!all(is.na(Tpref)), round(mean(Tpref, na.rm=T),2),  NA),
            Tbask=ifelse(!all(is.na(Tbask)), round(mean(Tbask, na.rm=T),2),  NA),
            Tforage_optim=ifelse(!all(is.na(Tforage_optim)), round(mean(Tforage_optim, na.rm=T),2), NA),
            Tforage_max=ifelse(!all(is.na(Tforage_max)), max(Tforage_max, na.rm=T), NA),  
            Tforage_min=ifelse(!all(is.na(Tforage_min)), min(Tforage_min, na.rm=T), NA), 
            Tmerge=ifelse(!all(is.na(Tmerge)), min(Tmerge, na.rm=T), NA),
            .groups='drop') %>%
  filter(!is.na(species)) # removing Hyla spp. and Hypopachus spp. traits
```
```{r, echo=F}
kable(tbif %>% slice_sample(n=3), digits = 3)
```
```{r}
full_join(tchar, tbif, by=c('species','lifeStage')) %>%
  mutate(refIDs=paste(refIDs_n, refIDs_a, sep=', ')) %>%
  select(-refIDs_a, -refIDs_n, -lifeStage) %>%
  select(species, everything()) %>%
  write.csv('../ATraiU2_summary_values_2022SEPT.csv', row.names = F)
```
```{r, echo=F}
kable(full_join(tchar, tbif, by=c('species','lifeStage')) %>%
  mutate(refIDs=paste(refIDs_n, refIDs_a, sep=', ')) %>%
  select(-refIDs_a, -refIDs_n, -lifeStage) %>%
  select(species, everything()) %>% 
    filter(species %in% c("Pseudacris crucifer","Lithobates clamitans","Anaxyrus fowleri")),
  digits=2)
````

# Double Checked Stats -----

```{r}
trait_long %>% 
  group_by(checked) %>%
  summarize(n=n(), .groups='drop') %>% mutate(per_n=round((n/sum(n)*100),1))
```

# Outlier Analysis for QA/QC -----

I checked for numerical outliers as a way to QA/QC the data. A numerical outlier is defined as a value with a z score > 2.49 at either a species, family, or entire trait scale. These numerical outliers might be used with caution for future database users. This occurs after two people double checked data entry. 

```{r}
outliers<-bind_rows(
  # all taxa for each trait
trait_long %>%
  filter(traitName != 'Activity', traitValue !='overwinters') %>%
  group_by(traitName, lifeStage) %>%
  mutate(tV=as.numeric(traitValue),
         z_score=(mean(tV)-tV)/sd(tV)) %>%
  select(species, traitName, traitValue, tV, z_score, traitSource) %>%
  filter(abs(z_score) > 2.49) %>%
  mutate(outlier = 'trait') %>% ungroup(),
# at the family level
trait_long %>%
  filter(traitName != 'Activity', traitValue !='overwinters') %>%
  group_by(traitName, family, lifeStage) %>%
  mutate(tV=as.numeric(traitValue),
         z_score=(mean(tV)-tV)/sd(tV)) %>%
  select(species, traitName, traitValue, tV, z_score, traitSource) %>%
  filter(abs(z_score) > 2.49) %>% ungroup() %>%
  mutate(outlier = 'family'),
# at the genus level
trait_long %>%
  filter(traitName != 'Activity', traitValue !='overwinters') %>%
  group_by(traitName, genus, lifeStage) %>%
  mutate(tV=as.numeric(traitValue),
         z_score=(mean(tV)-tV)/sd(tV)) %>%
  select(species, traitName, traitValue, tV, z_score, traitSource) %>%
  filter(abs(z_score) > 2.49) %>% ungroup() %>%
  mutate(outlier = 'genus'),
# at the species level
trait_long %>% ungroup() %>%
  filter(traitName != 'Activity', traitValue !='overwinters') %>%
  group_by(traitName, species) %>%
  mutate(tV=as.numeric(traitValue),
         z_score=(mean(tV)-tV)/sd(tV)) %>%
  select(species, traitName, traitValue, tV, z_score, traitSource) %>%
  filter(abs(z_score) > 2.49) %>%
  mutate(outlier = 'species'))
```

The above code outputs `r nrow(outliers %>% distinct(species, traitName, traitValue))` distinct traits to check, which represents `r round(nrow(outliers %>% distinct(species, traitName, traitValue))/ nrow(trait_long %>% filter(traitName !='Activity'))*100, 3)` % of the entire trait database. 

```{r, echo=F}
kable(outliers %>%
       group_by(species, traitName, traitValue) %>%
       summarize(z=paste(round(z_score,1), collapse=', '),
                 outs=paste(outlier, collapse=', '),
                 source=paste(unique(traitSource), collapse=', ')))
```

Each numerical outlier is checked for transcription error [last done on 03/22/2022]. Finally, I added the outliers flag to the full database and output the data in long format.

```{r}
trait_long %>%
  # bringing in the outliers flag
  left_join(outliers %>%
              group_by(species, traitName, traitValue) %>%
              summarize(outlier=paste(outlier, collapse=', '), .groups='drop'),
            by=c('species','traitName','traitValue')) %>%
  select(-checked, -dataCollationType) %>% 
  # fixing traits that have missing unit values
  mutate(traitUnit = case_when(is.na(traitUnit) & 
                                       traitName %in% c('Tpref','Tmerge','CTmax') & 
                                       traitValue != 'overwinters' ~ 'C', 
                         is.na(traitUnit) & traitName == 'Mass' ~ 'g', 
                         T ~ traitUnit)) %>%
  # add in the ITIS TSNs
  mutate(itisTSN=case_when(species == "Lithobates okaloosae" ~ 775103, species == "Lithobates sevosus" ~ 775113,
                           species == "Lithobates sylvaticus" ~ 775117, species == "Lithobates kauffeldi" ~ 1100485,
                           species == "Lithobates heckscheri" ~ 775092, species == "Lithobates capito" ~ 775083, 
                           species == "Lithobates virgatipes" ~ 775123, species == "Lithobates grylio" ~ 775091, 
                           species == "Lithobates areolatus" ~ 775078, species == "Lithobates palustris" ~ 775107,
                           species == "Lithobates sphenocephalus" ~ 775116, 
                           species == "Lithobates clamitans" ~ 775087, species == "Lithobates catesbeianus" ~775084,
                           species == "Pseudacris ornata" ~ 173531, species == "Pseudacris kalmi" ~ 774547,
                           species == "Pseudacris brimleyi" ~ 173524, species == "Pseudacris nigrita" ~ 173530, 
                           species == "Pseudacris fouquettei" ~ 774545, species == "Pseudacris ocularis" ~ 1105233, 
                           species == "Pseudacris brachyphona" ~ 173528, species == "Pseudacris feriarum" ~ 207307, 
                           species == "Pseudacris crucifer" ~ 207303, species == "Acris gryllus" ~ 173518, 
                           species == "Acris crepitans" ~ 173520, species == "Acris blanchardi" ~ 774220,
                           species == "Dryophytes avivoca" ~ 1098486, species == "Dryophytes femoralis" ~ 1095454,
                           species == "Dryophytes andersonii"~ 1098471, species == "Dryophytes gratiosus" ~ 1098493, 
                           species == "Dryophytes squirellus" ~ 1098561, 
                           species == "Dryophytes chrysoscelis" ~ 1098527,
                           species == "Dryophytes cinereus" ~ 1095381, species == "Dryophytes versicolor" ~ 1095495,
                           species == "Anaxyrus quercicus" ~ 773528, species == "Anaxyrus terrestris" ~ 773531, 
                           species == "Anaxyrus fowleri" ~ 773520, species == "Anaxyrus americanus" ~ 773511, 
                           species == "Incilius nebulifer" ~ 773583, species == "Scaphiopus holbrookii" ~ 173426, 
                           species == "Gastrophryne carolinensis" ~ 173467)) %>%
  #filter(is.na(itisTSN))
  #filter(traitName == 'CTmax', species == 'Gastrophryne carolinensis')
  #filter(is.na(traitUnit), traitName != 'Activity') 
  write_csv('../ATraiU2_full_2022SEPT.csv')
```
