==> ATraiU 2.0/code/2_Describing_ATraiU2.Rmd <==
---
title: "Quantifying results of Anuran Physiological Trait Database Search"
author: "TP DuBose"
date: "06/08/2022"
output: pdf_document
---

```{r setup, include=F}
library(tidyverse); library(cowplot); library(grid); library(knitr)
library(maps);library(sf)
```

The purpose of this code is to describe the physiological trait database for the manuscript results. We start by bringing in the necessary trait table and references. We also identify our entire focal species list.

```{r}
trait_long<-read_csv('../ATraiU2_full_2022SEPT.csv', show_col_types = F) %>%
  distinct(traitName, species, refID, traitValue, .keep_all=T) %>%
  filter(traitName != 'Tforage_optim')
refs <- read_csv('../ATraiU2_reference_list_2022SEPT.csv', show_col_types = F) 
foc_taxa<-c(trait_long %>% pull(species) %>% unique(), 
            "Pseudacris brimleyi", "Pseudacris nigrita", "Pseudacris ocularis")
```

Our focal taxa are `r foc_taxa`. 

We found trait values for the following traits: `r trait_long %>% pull(traitName) %>% unique()` .

Next we describe the results of this literature search. Our database contains `r nrow(trait_long)` reported trait values from `r trait_long %>% pull(refID) %>% unique() %>% length()` unique sources. Of those sources, `r refs %>% filter(refID %in% unique(trait_long$refID), grepl('al Review', journal)) %>% nrow()` are from the Herpetological Review. Below I explore the characteristics of the sources used to build this database.

```{r source info, echo=F, results='asis'}
kable(refs %>% distinct(refID, .keep_all=T) %>%
  select(ends_with('author')) %>%
  pivot_longer(everything()) %>% filter(!is.na(value)) %>%
  count(value) %>% # arrange(desc(n))
  summarize(n_total=nrow(.),
            n_mul4=sum(n>3),
            n_med=median(n),
            n_high=max(n),
            n_high_who=paste(ifelse(n==n_high,value,''), collapse='')),
  caption='Author information from sources in the database')
kable(refs %>% filter(refID %in% trait_long$refID)%>%
        distinct(refID, .keep_all=T) %>%
  count(journal) %>% arrange(desc(n)) %>% filter(n>10), 
  caption='Journals with more than 10 sources found within this database')
```

We then explore the timing and monotonic trends in publication dates within our sources. We remove the `r refs %>% filter(year < 1944) %>% nrow()` earliest years because it is so sparse: `r refs %>% filter(year < 1944) %>% pull(year)`

```{r}
library(Kendall)
n_t_year<-trait_long %>% filter(lifeStage %in% c('adult','unknown')) %>% 
  distinct(species, traitName, refID) %>%
  left_join(refs, by='refID') %>%
  count(year) %>%
  right_join(data.frame(year=min(refs$year):2022), by='year') %>%
  replace_na(list(n=0)) %>%
  arrange(year) %>% filter(year > 1944)
(mk_year<-MannKendall(n_t_year$n))
```
```{r time graph, echo=F, fig.height=2.5, fig.width=4}
n_t_year %>% ggplot()+geom_point(aes(x=year, y=n))+
  geom_text(data=. %>% filter(max(n)==n),
            aes(label=year, x=year+3, y=n))+
  geom_smooth(aes(x=year, y=n),method='loess',formula='y~x')+theme_classic()
```

# Quantifying database composition through trait coverage and trait completeness

We split traits into two types: physiological and behavioral. Below is the split of unique trait values for each. Unique trait values refer to observations that have a unique source, trait type (e.g., Mass), species, and life stage.

```{r, echo=F}
kable(trait_long %>%
  filter(species %in% foc_taxa) %>%
  mutate(type=ifelse(traitName %in% c("Mass","CTmax","CTmin","Tpref"),
                     'phys','behavior')) %>%
  group_by(type) %>%
  summarize(n=n(),
         n_species=n_distinct(species)) %>%
  ungroup() %>% mutate(per=n/sum(n)),
  caption='Total unique trait values for physiological and behavioral traits')
```

```{r, echo=F, results='asis'}
kable(trait_long %>% 
  group_by(traitName, traitEvidenceCategory) %>% 
  count() %>% 
  group_by(traitName) %>% mutate(total_n=sum(n)) %>% 
  filter(n==max(n)) %>%
  mutate(per.top=round(n/total_n*100, 1)) %>%
  arrange(traitEvidenceCategory),
  caption='Number of unique traits and top evidence category for each trait, across all taxa.')
```

## Trait Coverage

```{r}
kable(traitcov<-trait_long %>% 
  filter(species %in% foc_taxa, 
         lifeStage %in% c('adult','unknown')) %>%
  distinct(species, traitName) %>%
  group_by(traitName) %>%
  count() %>% arrange(desc(n)),
  caption='Trait coverage or number of species with atleaset one value for that traitName')
```

For trait coverage, we have a median of `r median(traitcov$n)` species for each trait (IQR = `r IQR(traitcov$n)` species). This is a median trait coverage of `r round(100*median(traitcov$n)/40, 1)` %. 

The following species have Tpref values: `r trait_long %>% filter(species %in% foc_taxa,          lifeStage %in% c('adult','unknown'), traitName %in% c('Tbask')) %>% pull(species) %>% unique()` .

# Trait completeness

We then built a table that contains the trait completeness (as number of traits) for the adult stage and all life stages. Life stages are binned into 5 groups: eggs, tadpoles, metamorphs, juveniles, and adults.

```{r}
ls_trait_total<-trait_long %>%
  filter(species %in% foc_taxa) %>%
  mutate(lifeStage = case_when(grepl('stage', lifeStage) ~ 'tadpole',
                                  lifeStage=='unknown' ~ 'adult',
                                  T ~ lifeStage)) %>%
  distinct(species, traitName, lifeStage, .keep_all=T) %>%
  group_by(species) %>%
  tally() %>% ungroup() %>% rename(n_LS=n)
trait_total<-trait_long %>%
  filter(species %in% foc_taxa, 
         lifeStage %in% c('adult','unknown')) %>%
  distinct(species, traitName, .keep_all=T) %>%
  group_by(species) %>%
  tally() %>% ungroup() %>%
  arrange(desc(n)) %>%
  left_join(ls_trait_total, by='species')
```

The following species have 0 trait values within our database: `r (mis_taxa<-foc_taxa[which(!(foc_taxa %in% trait_total$species), useNames = T)])`.

Below we explore the traits for these species with the fewest trait values. 

```{r}
trait_long %>% filter(lifeStage%in% c('adult','unknown')) %>%
  filter(species %in% trait_total[trait_total$n==min(trait_total$n),'species']) %>%
  select(species, traitName, lifeStage, traitSource)
```


```{r, echo=F}
kable(trait_total, caption='Trait completeness per focal species. n is trait completeness (out of 9 possible) for adults and n_LS is trait completeness across the 5 life stages (egg, tadpole, metamorph, juvenile, adult).')
```

Matrix completion describes the completeness of the matrix (so at least one trait value for each species and each trait; or 9 x 40 = 360 values). By summing the n column of the table above, we get `r sum(trait_total$n)` trait values and `r round(100*sum(trait_total$n)/360, 2)`% matrix completion. 

Thus in total, for the adult trait values, we found a median of `r trait_total %>% summarize(median(n)) %>% as.numeric()` (IQR = `r trait_total %>% summarize(IQR(n)) %>% as.numeric()`) reported trait values per species for the adult life stage. That translates to `r round(100*median(trait_total$n)/9)`% trait completeness. We found a median of `r ls_trait_total %>% summarize(median(n_LS)) %>% as.numeric()` (IQR = `r ls_trait_total %>% summarize(IQR(n_LS)) %>% as.numeric()`) reported trait values per species across life stage. The following species in our focal taxa had the most and minimum trait completeness for the adult (or unknown) lifestage. 

```{r, echo=F}
kable(trait_total %>% filter(n==max(n)|n==min(n)))
```

We then investigated the distribution of reported trait values across life stages in the table below. Of those considered adult in this analysis, `r trait_long %>% filter(lifeStage=='unknown') %>% nrow()` were not clearly reported. 

```{r, echo=F}
kable(trait_long %>%
  mutate(lifeStage = case_when(grepl('stage', lifeStage) ~ 'tadpole',
                                  lifeStage=='unknown' ~ 'adult',
                               lifeStage=='egg' ~ 'eggs',
                                  T ~ lifeStage)) %>%
  group_by(lifeStage) %>%
  summarize(n=n(), n_spp=n_distinct(species)) %>% arrange(desc(n)))
```

```{r}
trait_total %>% filter(max(n_LS)==n_LS) %>% pull(species)
trait_long %>%
  filter(species == (trait_total %>% filter(max(n_LS)==n_LS) %>% pull(species))) %>%
  mutate(lifeStage=case_when(lifeStage == 'unknown' ~ 'adult',
                             grepl('stage', lifeStage) ~'tadpole',
                             T ~ lifeStage)) %>%
  group_by(lifeStage) %>% summarize(n_distinct(traitName))

```

I use the RCS data for the area of occupancy and for conservation status. *Lithobates sevosus* was not included in the species pool for the RCS analysis. So I will use the IUCN range map to identify which HUC12 are occupied and calculate its area of occupancy (using the same data inputs as DuBose et al. 2022. I also looked up its conservation status (07/01/2022).

```{r, eval=F}
huc12<-readRDS('/home/tracidubose/huc12_wgs84_sf_02010323.rds')
#IUCN range maps downloaded 10/2020
iucn_lsev<-read_sf('/home/tracidubose/RCS_Anuran_input/ANURA/', 'ANURA')  %>%
 filter(grepl('sevosus',binomial), # only range of Lithobates sevosus
        grepl('Extant',legend)) # only the extant range
# make both shapefiles in same crs and do a spatial join
huc_keep<-iucn_lsev %>% st_transform(st_crs(huc12)) %>% st_join(huc12)
# sum the area of the watersheds that touch/overlap this range
huc_keep %>% distinct(huc12, .keep_all = T) %>%
  pull(WSAREA) %>% sum()
```

```{r rcs data}
# bringing in the RCS data to calculate area of occupancy
rcs<-read_csv('C:/Users/Owner/Desktop/Git/National-RCS/Anuran_RCS_and_components.csv',
              show_col_types=F) %>% 
  filter(grepl('North', extent)) %>%
  select(species, genus, family, IUCNcategory, ESA.Listing.Status, NatSGCN, AOO_WS_sqkm, RCS_WS) %>%
  # adding in Lithobates servosus
  bind_rows(data.frame(species='Lithobates sevosus',
                       genus='Lithobates', family='Ranidae',
                       IUCNcategory='CE', ESA.Listing.Status='Endangered',
                       NatSGCN = 'in =1 SWAP',
                       AOO_WS_sqkm=737.8))
```

And finally, we investigated how conservation status might influence trait completeness. Below we investigate which species in our focal taxa group are of conservation concern for the IUCN, ESA, and for the National Species of Greatest Conservation Need list.  

```{r, echo=F}
spocc<-rcs %>% filter(species %in% foc_taxa, 
                      IUCNcategory %in% c('NT', 'VU', 'EN','CE')|
                        ESA.Listing.Status !='Not Listed') %>%
  select(species, IUCNcategory, ESA.Listing.Status, NatSGCN) 
kable(spocc, 
      col.names = c('Species','IUCN status','ESA status', 'SGCN status'))
```

```{r, echo=F}
kable(trait_total %>% filter(species %in% spocc$species),
  caption='N trait completeness across life stages for species of conservation concern')
```

We can use a binomial test to see if this trait completeness is less than expected (given a 50% probability of a trait being reported in the literature).

```{r}
# Lithobates areolatus
binom.test(4,9)
# Lithobates capito, Lithobates sevosus
binom.test(2,9)
```

```{r, echo=F, fig.height=2, fig.width=4}
library(ggrepel)
trait_total %>%
  ggplot()+
  geom_histogram(aes(x=n), binwidth = 1)+
  geom_vline(data=. %>% filter(species %in% spocc$species), 
             aes(xintercept=n),
             linetype='dashed', size=1)+
  geom_text_repel(data=. %>% filter(species %in% spocc$species), 
            aes(x=n+0.1, y=5.3, label=species))+
  scale_x_continuous('N Traits',
                     breaks = seq(0,9,1),
                     expand = c(0,0))+
  scale_y_continuous('N species')+
  theme_classic()
```

# Figures

## Summary Figure 
### Trait Available Grid
```{r}
library(gridExtra);library(MetBrewer)
trait_order<-trait_long %>%
  filter(species %in% foc_taxa,
         traitName != 'Tforage_optim') %>%
  group_by(species, traitName) %>% slice(1) %>%
  group_by(traitName) %>% count() %>%
  arrange(desc(n)) %>%
  pull(traitName)
oc<-trait_long %>%
  filter(species %in% foc_taxa,
         traitName != 'Tforage_optim',
         lifeStage %in% c('adult','unknown')) %>%
  distinct(species, family, traitName, refID) %>%
  group_by(species, family, traitName) %>%
  summarize(n=n(),
         refss=paste(refID, collapse=', '),
         FamI=substr(family,1,1),
         TraF=factor(traitName, levels=trait_order,
                     labels=c('Activity','Mass',
                              expression(T[pref]),
                              'CT[max]',
                              'CT[min]',
                              'Tmerge',
                              'Tforage[max]',
                              'Tforage[min]',
                              'T[bask]')),
         Taxa=ifelse(species %in% spocc$species, paste0(species,'*'),species),
         .groups='drop') %>%
  ggplot()+
  geom_tile(aes(x=TraF, y=Taxa, fill=n))+
  scale_fill_gradient('Reference\nCount',
                       low='grey90', high='grey20',
                      breaks=c(1,4,8,12,16))+
  scale_x_discrete('Trait Name')+
  facet_wrap(~FamI, scales='free_y', ncol=1, strip.position = 'left')+
  theme_cowplot()+
  theme(strip.placement='outside',
        strip.background = element_rect(fill=NA),
        strip.text=element_blank(),#element_text(size=10),
        #panel.grid.major.y = element_line(color="black"),
        legend.justification = 'left',
        axis.text.y=element_text(size=8.5, face='italic'),
        legend.box.just = 'left',
        legend.text = element_text(size=8.5),
        legend.title = element_text(size=9),
        axis.text.x=element_text(angle=25,hjust=.92,size=8.5),
        axis.title.y = element_blank(),
        axis.title.x=element_text(size=9),
        plot.title = element_text(size=10, hjust=0.5))
ocp = ggplot_gtable(ggplot_build(oc))
#gtable::gtable_show_layout(ohp)
# get the number of unique x-axis values per facet (1 & 3, in this case)
y.var <- sapply(ggplot_build(oc)$layout$panel_scales_y,
                function(l) length(l$range$range))
# change the relative widths of the facet columns based on
# how many unique x-axis values are in each facet
ocp$heights[ocp$layout$t[grepl("panel", ocp$layout$name)]] <- ocp$heights[ocp$layout$t[grepl("panel", ocp$layout$name)]] * y.var
grid.draw(ocp)
ggsave('../manuscript/summary_grid_figure.jpg',
       width = 3.5, height = 5)
```

```{r, echo=F, fig.width=3.5, fig.height=6}
states<-st_as_sf(map('state', fill=T, plot=F))
usa<-st_as_sf(map('usa', fill=T, plot=F))
SE_states<-c('mississippi', 'georgia',
             'florida','south carolina', 'north carolina', 'virginia',
             'alabama', 'tennessee')
se_inset<-ggplot()+
  geom_sf(data=usa, fill='white')+
  geom_sf(data=states %>% filter(ID %in% SE_states),
          fill='lightgrey', color='lightgrey')+
  theme_void()

trait_ref_ndf<-trait_long %>%
  filter(species %in% foc_taxa,
         lifeStage %in% c('adult','unknown')) %>%
  group_by(species, family, traitName, refID) %>%
  slice(1) %>%
  group_by(species, family, traitName) %>%
  summarize(n=n(),
            refss=paste(refID, collapse=', '),
            Taxa=ifelse(species %in% spocc$species,
                        paste0(species,'*'),species),
            .groups='drop')  %>%
  distinct(.keep_all = T)
trait_ref_ndf %>% group_by(species) %>% summarize(tn=sum(n)) %>% pull(tn) %>% max()
tax_ref_ord<-trait_ref_ndf %>%
  group_by(Taxa) %>% summarize(tn=sum(n)) %>%
  arrange(desc(tn)) %>%
  pull(Taxa)
trait_hist<-trait_ref_ndf %>%
  mutate(TaxF=factor(Taxa, levels=rev(tax_ref_ord)),
         TraitF=factor(traitName, 
                       levels=c('Mass','CTmax','CTmin','Tpref',
                                'Tbask','Tforage_min','Tforage_max',
                                'Tmerge','Activity'))) %>%
  ggplot()+
  geom_col(aes(y=TaxF, x=n, fill=TraitF), position='stack')+
  draw_plot(se_inset, x=27, y=17.5, scale=28)+
  labs(y='',x='Reference count')+
  scale_fill_manual('Trait Name', values=met.brewer('Hiroshige')[-5])+
  scale_x_continuous(breaks=seq(0,40,5), 
                     expand=c(0.01,0.01), limits = c(0,40))+
  theme_classic()+
  theme(legend.position=c(.735,.20), 
        legend.text=element_text(size=9),
        legend.title=element_text(size=9),
        legend.key.size = unit(1, "lines"),
        #legend.background = element_rect(fill=NA),
        axis.text.y = element_text(face='italic'),
        axis.title.y=element_blank(),
        panel.grid.major.x = element_line(color='grey90'))
trait_hist
ggsave('../manuscript/F1_nref.jpg',
       width=3.5, height=6)
```
```{r, eval=F}
ggsave(plot=grid.arrange(ocp,trait_hist, nrow=1),
       '../manuscript/summary_figure_combo.jpg',
       width=8, height=6)
```

## Comparison among taxonomic groups

I need to add the species with no traits to our trait table. 

```{r}
trait_count_rarity<-data.frame(species=foc_taxa) %>%
  filter(!(species %in% trait_total$species)) %>%
  mutate(n=0, n_LS=0)%>%
  bind_rows(trait_total) %>%
  left_join(rcs, by='species')
nrow(trait_count_rarity)
```
### Comparing Trait Completeness among Families

```{r family table, echo=F}
kable(trait_count_rarity %>%
  group_by(family) %>%
  summarize(n_sp=n(),
            med_trait_n=median(n),
            iqr_traits=IQR(n),
            range_traits=paste(min(n), max(n), sep=' to ')) %>%
    arrange(med_trait_n),
  caption='Family level information about trait completeness')
```

```{r}
set.seed(12345)
kruskal.test(n~family, data=trait_count_rarity)
```

I downloaded phylopics that represent each family. Below are people I need to acknowledge for these pictures. 

```{r}
# getting phylopic information
# tried to use r package api but couldn't get it to work past examples

# Hyla versicolor - Will Booker
hyla <- "../manuscript/hyla.png"
# Incilius valliceps - Jose Carlos Arenas-Monroy
incilius<-"../manuscript/incilius.png"
# Lithobates catesbeianus - Chuanixn Yu
lithobates<-"../manuscript/lithobates.png"
# Anaxyrus woodhousii - Michael Scroggie
anaxyrus<-"../manuscript/anaxyrus.png"
# Pseudacris maculata - Beth Reinke
pseudacris<-"../manuscript/pseudacris.png"
# Hypopachus variolosus - Jose Carlos Arenas-Monroy
microhyla<-"../manuscript/gastrophryne.png"
# Spea multiplicata - Jose Carlos Arenas-Monroy
spea<-"../manuscript/scaphiopus.png"
library(magick)
```

```{r, echo=F}
gen_pal<-met.brewer('Veronese', n=length(unique(trait_count_rarity$family)),
                    type='discrete')
fam_ord<-trait_count_rarity %>%
  group_by(family) %>%
  summarize(n_sp=n(),
            med_trait_n=median(n)) %>%
  arrange(desc(med_trait_n)) %>%
  pull(family)

fam_plot<-trait_count_rarity %>%
  mutate(famF=factor(family, levels=rev(fam_ord))) %>%
ggplot(aes(x=n, y=famF, fill=family))+
  geom_boxplot(#data=. %>% 
                 # don't want lines for family with only 1 taxa
                # filter(!(family %in% c('Scaphiopodidae', 'Microhylidae'))),
               aes(color=family), alpha=0.5, outlier.alpha=0)+
  geom_point(position=position_jitter(height = .3, width=0),
             pch=1, color='black', size=2)+
  #geom_text(data=. %>% group_by(famF) %>% count(),
  #          aes(x=9.3,label=n), color='black')+
  draw_image(hyla,  x = 8.5, y = 0.5, scale = 1)+
  draw_image(lithobates,  x = 8.5, y = 1.5, scale = 1)+
  draw_image(microhyla,  x = 8.5, y = 2.5, scale = 1)+
  draw_image(anaxyrus,  x = 8.5, y = 3.5, scale = 1)+
  draw_image(spea,  x = 8.5, y = 4.5, scale = 1)+
  labs(y='Family')+
  scale_x_continuous('Trait Completeness', breaks=seq(0,8,2),
                     limits = c(0,9))+
  scale_color_manual('Family', values=gen_pal, 
                     aesthetics = c('color','fill'))+
  theme_classic()+
  theme(legend.position='none')
fam_plot
```

### Comparison of Trait completeness among genera

```{r genera table, echo=F}
kable(trait_count_rarity %>%
  group_by(genus) %>%
  summarize(n_sp=n(),
            med_trait_n=median(n),
            iqr_traits=IQR(n),
            range_traits=paste(min(n), max(n), sep=' to ')) %>%
    arrange(med_trait_n), 
  caption='Differences among genera in trait completeness')
```
```{r}
kruskal.test(n~genus, data=trait_count_rarity)
library(dunn.test)
dunn.test(trait_count_rarity$n,
          trait_count_rarity$genus,
          method="bonferroni", alpha=0.05)
```

```{r, echo=F}
gen_ord<-trait_count_rarity %>%
  group_by(genus) %>%
  summarize(n_sp=n(),
            med_trait_n=median(n)) %>%
  arrange(desc(med_trait_n)) %>%
  pull(genus)
gen_plot<-trait_count_rarity %>%
  mutate(genF=factor(genus, levels=rev(gen_ord))) %>%
  ggplot(aes(x=n, y=genF, fill=family))+
  geom_boxplot(aes(color=family), alpha=0.5, outlier.alpha=0)+
   geom_point(position=position_jitter(height = .3, width=0),
             size=2, color='black', pch=1)+
  labs(y='Genus')+
  scale_x_continuous('Trait Completeness', breaks=seq(0,8,2))+
  scale_color_manual('Family',values=gen_pal,
                     aesthetics=c('color','fill'))+
  theme_classic()+
  theme(legend.text = element_text(size=8),
        legend.title = element_text(size=8),
        legend.background = element_rect(fill=NA, color=NA),
        legend.position='bottom')

library(cowplot)
legend_b <- get_legend(gen_plot)
ggsave(plot=plot_grid(gen_plot+theme(legend.position='none'),
                      legend_b, ncol=1, rel_heights = c(1,.1)),
       '../manuscript/genera_only_plot.jpg', width=3.5, height=3)


plot_grid(plot_grid(gen_plot+theme(legend.position='none', axis.title = element_text(size=8)), 
                    fam_plot+theme(axis.title = element_text(size=8)), labels='AUTO'), 
          legend_b, ncol = 1, rel_heights = c(1.3, .1))
ggsave('../manuscript/F2_tax_comp_wtax.jpg', 
       width = 6.5, height = 2.5)
```

## Compare Rarity with Trait Completeness

```{r}
cor.test(trait_count_rarity$AOO_WS_sqkm,
         trait_count_rarity$n, method='spearman', exact=F)
```
```{r, echo=F, fig.width=2.5, fig.height=2.5}
library(scales)
tcor_plot<-trait_count_rarity %>%
  ggplot()+
  geom_point(aes(x=log10(AOO_WS_sqkm), y=n), pch=1)+
  scale_x_continuous(expression('Area of Occupancy, km '^2),
                     breaks=c(3.30103,4,4.69897,5.69897),
                     labels = function(x){comma(10^x)},
                     limits = c(3.30103, 5.77))+
  scale_y_continuous('Trait Completeness')+
  theme_classic()+
  theme(axis.text.x = element_text(angle=20, hjust=.9),
        axis.title = element_text(size=9))
tcor_plot
```

I recount the number of traits at each life stage so I can keep track of the type and number of life stages. 

```{r}
trait_ls_total<- trait_long %>%
  filter(species %in% foc_taxa) %>%
  mutate(lifeStage = case_when(grepl('stage', lifeStage) ~ 'tadpole',
                                  lifeStage=='unknown' ~ 'adult',
                                  T ~ lifeStage)) %>%
  distinct(species, traitName, lifeStage, .keep_all=T) %>%
  group_by(species) %>%
  arrange(lifeStage) %>%
  mutate(l_abrev=toupper(substr(lifeStage, 1,1)))%>%
  summarize(n=n(),
            total_ls=paste(unique(l_abrev), collapse=', '),
            n_ls=n_distinct(l_abrev)) %>% 
  ungroup() %>%
  arrange(desc(n)) %>% 
  bind_rows(data.frame(species=foc_taxa) %>%  mutate(n=0)) %>%
  group_by(species) %>% arrange(desc(n)) %>% slice(1) %>%
  left_join(rcs, by='species')
```

```{r, echo=F}
kable(head(trait_ls_total %>% select(species, n, total_ls, n_ls)))
```

The potential combinations of life stages are: `r unique(trait_ls_total$total_ls)`. 

Based on the trait data, I think total possible trait completeness for all life stages would be `r 9+2+4+3+5`.
```{r, echo=F, results='asis'}
kable(trait_long %>% 
        filter(species %in% foc_taxa) %>%
        mutate(lifeStage = case_when(grepl('stage', lifeStage) ~ 'tadpole',
                                  lifeStage=='unknown' ~ 'adult',
                                  T ~ lifeStage)) %>%
      group_by(lifeStage) %>% 
        summarize(`trait coverage`= n_distinct(traitName), 
                  traits = paste(unique(traitName), collapse=', ')))
```

```{r}
cor.test(trait_ls_total$n, trait_ls_total$AOO_WS_sqkm, 
         method='spearman', exact=F)
```
```{r, echo=F}
ls_cor_plot<-trait_ls_total %>%
  replace_na(list(n_ls=0))%>%
  ggplot()+
  geom_point(aes(x=log10(AOO_WS_sqkm), y=n, fill=as.character(n_ls),
                 color=as.character(n_ls), shape=as.character(n_ls)),
             size=2, alpha=0.75)+
  geom_text_repel(data=. %>% ungroup() %>% 
                    filter(max(n)==n) %>%
                    mutate(species=paste0("italic('",species,"')")),
                  aes(x=log10(AOO_WS_sqkm), y=n, label=species),
                  size=2.4, parse=T,
                  min.segment.length = 0)+
  scale_x_continuous(expression('Area of Occurrence, km '^2),
                     breaks=c(3.30103,4,4.69897,5.69897),
                     labels = function(x){comma(10^x)},
                     limits = c(3.30103, 5.77))+
  scale_y_continuous('Trait Completeness', breaks=seq(0,16,2))+
  scale_color_manual('Life\nStages',
                     values=c('grey70','grey50',
                              'grey40','grey20','grey10','black'),
                     aesthetics = c('color','fill'))+
  scale_shape_manual('Life\nStages',values=c(21,22,24,25,23,19))+
  theme_classic()+
  theme(#legend.position = c(.1,.7),
        legend.text = element_text(size=8),
        legend.title = element_text(size=8),
        axis.title = element_text(size=9),
        axis.text.x = element_text(angle=20, hjust=.9))
ls_cor_plot

plot_grid(tcor_plot, ls_cor_plot, labels='AUTO',rel_widths = c(.45,.55))
ggsave('../manuscript/F3_trait_aoo_corr.jpg', width=5, height=2.5)
```

Below I do some data manipulation to help describe life stage stuff. `r trait_ls_total %>% filter(grepl('T', total_ls)) %>% nrow()` species had information for the tadpole life stage available.

```{r, echo=F}
kable(trait_ls_total %>% filter(n_ls > 3) %>% select(species, n, total_ls),
      caption='Species with more than two life stages with trait values available')
kable(trait_ls_total %>% 
  mutate(n_ls_grp=ifelse(n_ls > 3 & !is.na(n_ls), 'more', 'less')) %>% 
  group_by(n_ls_grp) %>% summarize(min(AOO_WS_sqkm)),
  caption='minimum AOO of species with > 3 life stages with trait values reported or < 3 life stages with trait values reported')
```

# Stats used in the Discussion

```{r}
n_b4<-trait_long %>% left_join(refs, by='refID') %>%
  distinct(species, traitName, refID, .keep_all=T) %>%
  filter(year < 2008) %>% nrow()
n_all<-trait_long %>% left_join(refs, by='refID') %>%
  distinct(species, traitName, refID, .keep_all=T) %>%
  nrow()
```

The percent of all traits before 2008 (the year after Lemmon et al. 2007 clarified Pseudacris phylogeny) is `r round(n_b4/n_all*100, 1)`%.

```{r most numerous trait, echo=F}
mx_sp<-trait_long %>%
  filter(traitName !='Mass',
         traitName !='Activity',
         lifeStage %in% c('adult','unknown')) %>%
  distinct(species, traitName, verbatimLocality) %>%
  count(traitName, species) %>%
  filter(n > 3) %>% arrange(desc(n)) %>% filter(max(n)==n) %>% pull(species)
kable(trait_long %>% filter(traitName=='CTmax', 
                            species == mx_sp, 
                            lifeStage %in% c('adult','unknown')) %>% 
        select(species, traitName, traitValue, verbatimLocality),
      caption='Species with the highest number of values for a single trait type (CTmax)')
```
